changed security_deposit to float from integer

+++ migrations/20201014162935_users-table.js
@@ -18,7 +18,7 @@ exports.up = function(knex) {
            .notNullable();
         tbl.string('room_type', 128)
            .notNullable();
-        tbl.integer('security_deposit')
+        tbl.float('security_deposit')
             .notNullable();
         tbl.integer('guests_included')
            .notNullable();
